Finish BFS anim, move sidebar/canvas to component
